rearchitecture so that `useEffect` is not needed
